Allow users to switch charset encodings of APIC frame, fixes #42 (#43)
There are some quirks with APIC frame parsing on macOS iTunes and Finder. When the description is encoded in Unicode and is either an empty string or with an non-Western character as the last character, iTunes will fail to parse the picture from the frame.

This fix workaround it by always use Western encoding when the description is empty, and only opt to Unicode encoding when the caller explicitly choose to with `useUnicodeEncoding: true`. README.md is updated accordingly with a warning.